Refactor tests namespace to avoid possible conflict.
