+++ composer.json
@@ -38,7 +38,7 @@
   },
   "autoload-dev": {
     "psr-4": {
-      "Test\\": "tests/"
+      "Yajra\\Datatables\\Tests\\": "tests/"
     }
   },
   "extra": {
